Initial commit!  :confetti_ball:
This is a proof-of-concept of using TypeScript for a Sphere v2 game.
The game's Cellscript calls the TypeScript compiler programmatically to
convert any `.ts` sources in the project to `.js`.  Every attempt is
made to match the semantics of `tsc`, but there may still be some
glitches to work out.  Overall this first attempt seems to work well,
though.

Note that it is required to run `npm install typescript` first, as the
TypeScript compiler is not included.